feat(presentation/components): create card list file
